pyproject.toml
[project]
name = "notifications-mcp"
version = "0.1.0"
description = "MCP useful for notifications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
